refactor: simplify build and packaging of media-object
