+++ packages/components/media-object/bs-media-body.js
