+++ packages/components/media-object/css/index.js
@@ -1,0 +1,2 @@
+export { BsMediaCss } from './bs-media.css.js';
+export { BsMediaGroupCss } from './bs-media-group.css.js';
