+++ packages/components/media-object/index.js
@@ -1,5 +1,4 @@
+export * from './css'
 export { BsMedia } from './bs-media.js';
-export { BsMediaCss } from './bs-media.css.js';
 export { BsMediaBody } from './bs-media-body.js';
 export { BsMediaGroup } from './bs-media-group.js';
-export { BsMediaGroupCss } from './bs-media-group.css.js';
